Do PR refactoring
